Minerchest mod added to modpack.
New mod is added in this commit and it contains one element:
minerchest. The idea behind this self-organizing storage node
is to eliminate burden of manual combination of ingots and other
items mined and produced in large quantities. Putting such items
in this chest will automatically compact them into blocks.

Version changed to 2.8.

Signed-off-by: Michal Cieslakiewicz <michal.cieslakiewicz@wp.pl>